+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Unreleased (Use snapshot builds to get these changes)
   ``device.set_temperature_comfort()``, ``device.set_temperature_lowering()``,
   ``device.valve_state``
 * Manager: The manager can now be used with Python 3 (testing in progress)
+* Windows: Made CCU detection platform independent
 
 Incompatible (possible manual changes needed):